feat(tests): refactor and add tests to test service provider
